requirements.txt
Django~=5.2.1
djangorestframework~=3.16.0
django-cors-headers~=4.7.0
django-storages~=1.14.6
